umath refactor
